apps/ai-service/pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "AI Server for DS Bridge - Chat-based UI Builder"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "openai>=1.59.0",
    "anthropic>=0.43.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.0",
    "firebase-admin>=6.6.0",
    "python-multipart>=0.0.22",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.8.0",
    "mypy>=1.14.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
